+++ mlserver/repository/__init__.py
@@ -4,8 +4,11 @@ from .repository import (
     DEFAULT_MODEL_SETTINGS_FILENAME,
 )
 
+from .factory import ModelRepositoryFactory
+
 __all__ = [
     "ModelRepository",
     "SchemalessModelRepository",
     "DEFAULT_MODEL_SETTINGS_FILENAME",
+    "ModelRepositoryFactory",
 ]
